Fixed listing of ACL whitelist
A pattern I've discovered is that `array.forEach` calls seem to work
asynchronously, and `msg.send` calls that happen within them are emitted
before other `msg.send` calls in the same block. The trick, then, is to
aggregate the data to send, unshift the initial information to send to
the aggregated data, and then send in bulk.