Refactor TweetStream and fix unsub bug
Separated the twitter functionality into three additional files:

- twitter-types contains and returns the struct detailing the allowed
  follow types.
- twitter-stream contains and returns the `TwitterStream` class
  detailing a stream and what room it belongs to.
- twitter-tweetstream contains and returns the `TweetStream` class that
  has the functionality for subscribing and unsubscribing from streams.

The twitter script now just requires and instantiates a `TweetStream`,
and wires it to the robot.

The unsub bug was found in `unsubscribe`, which was incorrectly
assigning the return value of `_.remove` to the robot brain, instead of
the array from which items were removed. (`remove` returns an array of
_removed_ elements, while modifying the array passed to it.)