Working docker configuration!
This provides working docker configuration for the following:

- Building the nginx and hubot containers. The specific `Dockerfile`s
  are in `etc/docker/`.
- `docker-compose` configuration for development. Settings may be
  overridden using a `docker-compose.override.yml` file. Note that the
  `hubot` service uses the long-form `volumes` syntax; I found that the
  short-form was causing a conflict when mounting the volume otherwise.
- `docker-stack.yml` is for use with `docker stack`, and defines the
  production configuration. You then deploy using `docker stack deploy
  --compose-file=docker-stack.yml`. The primary difference is that it
  does not attempt to use local volumes, and uses tagged images always.

nginx is now configured to serve both HTTP and HTTPS, with HTTP
redirecting to HTTPS. HTTPS checks for paths matching `/github/`, and,
if found, performs a reverse proxy to the hubot; anything else serves
a dummy homepage.

Redis is configured to persist data to the `/data/` directory of the
host; this allows backups and persistence when the services are taken
down.