Fixed volume issue for redis
It feels like I'm going back and forth on this, but I suspect the issue
had to do with differences in running locally versus on the server.
Anyways, adding a fully qualified path and mapping it to the server
itself appears to have made a difference now that the deployment
environment works.