Listen and respond to "tweet" events
The `tweetstream` script now listens to `tweet` events, passing them on
to a new `Tweeter` class to process. `Tweeter` tweets the provided
status, using the ZF logo as a media attachment.

The github-release hook now emits the `tweet` event, with the message
`Released: #{release_name}: #{release_url}`.

To enable this, a new directory, `img`, now needs to be copied into the
container.