Ensures we only post once per Discourse watch
Some odd logic was causing duplication of watches stored in Redis, and
it appears it may have snuck in during the `restore()` operation.

This patch adds some logic to ensure no duplicate watches are created
during initialization; further, it always saves the current list of
watches to the brain once `restore()` completes to ensure we are
initialized with a known-good state.