Refactors bot to use listener middleware for authorizations
Actions that require authorization now add the listener metadata
`id: "authorize"`. A listener registered in `scripts/zf-acls.coffee`
then checks for that, and, when found, passes the message to the
`ZfAcl.verify` method. If that method returns a boolean false, a message
is sent indicating lack of authorization, and the response is marked
done; otherwise, it continues on to the next (which will be the actual
handler).

This approach allows removal of the redundant `lib/authorize.coffee`,
and removal of all code that called on it previously.