Use the "sender" object for user details of a pull request
Fixes #4

Per the [Webhooks Payloads documentation](https://developer.github.com/webhooks/#payloads),
each payload includes a "sender" object, and that object is the person
who initiated that particular event. In most cases, we can use the
`event.user` object, but for pull requests, it needs to be based on the
sender, as that person is the person:

- commenting on
- closing
- merging

the pull request.