Simplified image build and deployment
Adds a `Makefile` and several helper scripts to aid in building docker
images and deploying them to production.

The `Makefile` will create new images, tagging them with a value based
on the current date and time (e.g., `201804091145`). When you run `make
deploy`, it will find the latest tag for each image, and use that to
create a `docker-stack.yml` based on the `docker-stack.yml.dist` file,
and then deploy to the production swarm. Once deployed, it removes the
generated `docker-stack.yml` file.